% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/albums.R
\name{album_to_df}
\alias{album_to_df}
\title{Convert genius_album object to a data frame}
\usage{
album_to_df(x)
}
\arguments{
\item{x}{a \code{genius_album} object}
}
\value{
a tibble
}
\description{
Convert genius_album object to a data frame
}
\examples{
\dontrun{
album <- get_album(album_id = 337082)
album_to_df(album)
}

}
