% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{artist_meta}
\alias{artist_meta}
\title{Retrieve meta data for an artist}
\usage{
artist_meta(artist_id)
}
\arguments{
\item{artist_id}{An artist ID (\code{artist_id} returned in \code{\link{search_song}}).}
}
\description{
The Genius API lets you search for meta data for an artist, given an artist ID.
}
\examples{
artist_meta(artist_id = 129372)
}
