% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artists.R
\name{artist_to_df}
\alias{artist_to_df}
\title{Convert genius_artist object to a data frame}
\usage{
artist_to_df(x)
}
\arguments{
\item{x}{a \code{genius_artist} object}
}
\value{
a tibble
}
\description{
Convert genius_artist object to a data frame
}
\examples{
\dontrun{
artist <- get_artist(artist_id = 16775)
artist_to_df(artist)
}

}
