% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{browse_genius}
\alias{browse_genius}
\title{Open the Genius homepage in your browser}
\usage{
browse_genius()
}
\value{
A browser is opened to the Genius website
if the session is interactive. The URL is returned as a character string.
}
\description{
Opens a browser to \code{https://genius.com/}.
}
\examples{
\donttest{
browse_genius()
}
}
