% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/albums.R
\name{gen_get_album}
\alias{gen_get_album}
\title{Retrieve metadata for an album}
\usage{
gen_get_album(album_id, access_token = genius_token())
}
\arguments{
\item{album_id}{ID of the album (\code{album_id} within an object returned by
\code{\link{gen_get_song}})}

\item{access_token}{Genius' client access token, defaults to \code{genius_token}}
}
\value{
a list
}
\description{
The Genius API lets you request data for a specific album, given an album ID.
\code{gen_get_album()} returns this data in full.
}
\examples{
\dontrun{
gen_get_album(album_id = 337082)
}
}
\seealso{
[gen_get_album_df()] to return a tidy data frame.
}
