% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracklists.R
\name{gen_get_album_tracklist}
\alias{gen_get_album_tracklist}
\title{Retrieve an album's tracklisting}
\usage{
gen_get_album_tracklist(album_id, access_token = genius_token())
}
\arguments{
\item{album_id}{ID of the album (\code{album_id} within an object returned by
\code{\link{gen_get_song}})}

\item{access_token}{Genius' client access token, defaults to \code{genius_token}}
}
\value{
a tibble
}
\description{
Get an album's tracklisting, and song meta data, given an album ID.
}
\examples{
\dontrun{
gen_get_album_tracklist(album_id = 337082)
}
}
