% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/songs.R
\name{gen_get_song}
\alias{gen_get_song}
\title{Retrieve metadata for a song}
\usage{
gen_get_song(song_id, access_token = genius_token())
}
\arguments{
\item{song_id}{ID of the song (\code{song_id} within an object returned by
\code{\link{search_song}})}

\item{access_token}{Genius' client access token, defaults to \code{genius_token}}
}
\value{
a list
}
\description{
The Genius API lets you return data for a specific song, given a song ID.
\code{gen_get_song} returns this data in full.
}
\examples{
\dontrun{
gen_get_song(song_id = 3039923)
}
}
\seealso{
[gen_get_song_df()] to return a tidy data frame.
}
