% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{gen_search}
\alias{gen_search}
\title{Search documents hosted on Genius}
\usage{
gen_search(search_term, n_results = 10, access_token = genius_token())
}
\arguments{
\item{search_term}{A character string to search for}

\item{n_results}{Maximum no. of search results to return
(this is the number of hosted content search results, unique artist matches will
be smaller)}

\item{access_token}{Genius' client access token, defaults to \code{genius_token}}
}
\value{
a list
}
\description{
The Genius API lets you search hosted content (all songs). Use \code{gen_search()}
to return hits on for a given search term, in full.
}
\examples{
\dontrun{
search_artist(search_term = "Lil", n_results = 100)
}
}
