% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{genius_token}
\alias{genius_token}
\title{Get or set Genius access token value}
\usage{
genius_token(force = FALSE)
}
\arguments{
\item{force}{force setting a new Genius API token for the current environment?}
}
\value{
atomic character vector containing the Genius API token
}
\description{
The API wrapper functions in this package all rely on a Genius client access token
residing in the environment variable \code{GENIUS_API_TOKEN}. The
easiest way to accomplish this is to set it in the `\code{.Renviron}` file in your
home directory.
}
