% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geniusr.R
\docType{package}
\name{geniusr}
\alias{geniusr}
\alias{geniusr-package}
\title{\code{geniusr} package}
\description{
An interface to Genius' Web API.
}
\details{
See the README on
\href{https://github.com/ewenme/geniusr#readme}{GitHub}
}
