% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/albums.R
\name{get_album}
\alias{get_album}
\title{Retrieve metadata for an album}
\usage{
get_album(album_id, access_token = genius_token())
}
\arguments{
\item{album_id}{ID of the album (\code{album_id} within an object returned by
\code{\link{get_song_meta}})}

\item{access_token}{Genius' client access token, defaults to \code{genius_token}}
}
\description{
The Genius API lets you return data for a specific album, given an album ID.
\code{get_album} returns this data in a relatively untouched state.
}
\examples{
\dontrun{
get_album(album_id = 337082)
}
}
