% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{get_album_meta}
\alias{get_album_meta}
\title{Retrieve meta data for an album}
\usage{
get_album_meta(album_id)
}
\arguments{
\item{album_id}{An album ID (\code{album_id} returned in \code{\link{song_meta}}).}
}
\description{
The Genius API lets you search for meta data for an album, given an album ID.
}
\examples{
album_meta(album_id = 129372)
}
