% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/albums.R
\name{get_album_tracklist_search}
\alias{get_album_tracklist_search}
\title{Retrieve an album's tracklisting}
\usage{
get_album_tracklist_search(artist_name, album_name)
}
\arguments{
\item{artist_name}{Name of artist}

\item{album_name}{Name of album}
}
\value{
a tibble
}
\description{
Attempt to get an album's tracklisting, given an artist
and album name.
}
\examples{
\dontrun{
get_album_tracklist_search(artist_name = "Kendrick Lamar",
album_name = "DAMN.")
}
}
\seealso{
See \code{\link{get_album_tracklist_id}} to search for
an album tracklist using an album ID.

Other album: 
\code{\link{get_album_df}()},
\code{\link{get_album_tracklist_id}()},
\code{\link{get_album}()}
}
\concept{album}
