% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lyrics.R
\name{get_lyrics}
\alias{get_lyrics}
\title{Retrieve lyrics from a Genius lyrics URL}
\usage{
get_lyrics(song_url)
}
\arguments{
\item{song_url}{song lyric page URL (like in \code{song_url} returned by \code{\link{search_song}})}
}
\description{
Get lyrics from Genius' lyric pages.
}
\examples{
get_lyrics(song_url = "https://genius.com/Mike-will-made-it-california-rari-lyrics")
}
