% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lyrics.R
\name{get_lyrics_search}
\alias{get_lyrics_search}
\title{Retrieve lyrics associated with a Genius song}
\usage{
get_lyrics_search(artist_name, song_title)
}
\arguments{
\item{artist_name}{Name of artist}

\item{song_title}{Title of song}
}
\description{
Attempt to get lyrics from Genius' lyric pages using an associated
artist name and song title.
}
\examples{
\dontrun{
get_lyrics_search(artist_name = "Anderson .Paak",
song_title = "Come Home")
}
}
