% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lyrics.R
\name{get_lyrics_url}
\alias{get_lyrics_url}
\title{Retrieve lyrics associated with a Genius lyrics page URL}
\usage{
get_lyrics_url(song_lyrics_url)
}
\arguments{
\item{song_lyrics_url}{song lyrics url (like in \code{song_lyrics_url} returned by \code{\link{get_song_meta}})}
}
\description{
Scrape lyrics from a Genius' lyric page using it's associated URL. Best used with \code{\link{scrape_tracklist}}, when song IDs aren't returned - otherwise, \code{\link{scrape_lyrics_id}} is recommended.
}
\examples{
\dontrun{
get_lyrics_url(song_lyrics_url = "https://genius.com/Kendrick-lamar-dna-lyrics")
}
}
\seealso{
See \code{\link{get_lyrics_id}} to search lyrics using a song
ID, \code{\link{get_lyrics_search}} to search using artist name and song
title

Other lyrics: 
\code{\link{get_lyrics_id}()},
\code{\link{get_lyrics_search}()}
}
\concept{lyrics}
