% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/songs.R
\name{get_song}
\alias{get_song}
\title{Retrieve data for a song}
\usage{
get_song(song_id, access_token = genius_token())
}
\arguments{
\item{song_id}{ID of the song (\code{song_id} within an object returned by
\code{\link{search_song}})}

\item{access_token}{Genius' client access token, defaults to \code{genius_token}}
}
\description{
The Genius API lets you return data for a specific song, given a song ID.
\code{get_song} returns this data in a relatively untouched state.
}
\examples{
\dontrun{
get_song(song_id = 3039923)
}
}
