% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/albums.R
\name{print.genius_album}
\alias{print.genius_album}
\title{Slightly more human-readable output for genius_album objects}
\usage{
\method{print}{genius_album}(x, ...)
}
\arguments{
\item{x}{a genius_album object}

\item{...}{ignored}
}
\description{
Slightly more human-readable output for genius_album objects
}
