% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{print.genius_annotation}
\alias{print.genius_annotation}
\title{Slightly more human-readable output for genius_annotation objects}
\usage{
\method{print}{genius_annotation}(x, ...)
}
\arguments{
\item{x}{a genius_annotation object}

\item{...}{ignored}
}
\description{
Slightly more human-readable output for genius_annotation objects
}
