% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artists.R
\name{print.genius_artist}
\alias{print.genius_artist}
\title{Slightly more human-readable output for genius_artist objects}
\usage{
\method{print}{genius_artist}(x, ...)
}
\arguments{
\item{x}{a genius_artist object}

\item{...}{ignored}
}
\description{
Slightly more human-readable output for genius_artist objects
}
