% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{print.genius_referent}
\alias{print.genius_referent}
\title{Slightly more human-readable output for genius_referent objects}
\usage{
\method{print}{genius_referent}(x, ...)
}
\arguments{
\item{x}{a genius_referent object}

\item{...}{ignored}
}
\description{
Slightly more human-readable output for genius_referent objects
}
