% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/songs.R
\name{print.genius_song}
\alias{print.genius_song}
\title{Slightly more human-readable output for genius_song objects}
\usage{
\method{print}{genius_song}(x, ...)
}
\arguments{
\item{x}{a genius_song object}

\item{...}{ignored}
}
\description{
Slightly more human-readable output for genius_song objects
}
