% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lyrics.R
\name{scrape_lyrics}
\alias{scrape_lyrics}
\title{Retrieve lyrics assoicated with a Genius song ID}
\usage{
scrape_lyrics(song_id, access_token = genius_token())
}
\arguments{
\item{song_id}{song ID (like in \code{song_id} returned by \code{\link{search_song}})}

\item{access_token}{Genius' client access token, defaults to \code{genius_token}}
}
\description{
Scrape lyrics from Genius' lyric pages.
}
\examples{
scrape_lyrics(song_id = 3214267)
}
