% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracklist.R
\name{scrape_tracklist}
\alias{scrape_tracklist}
\title{Retrieve an album's tracklisting}
\usage{
scrape_tracklist(album_id, access_token = genius_token())
}
\arguments{
\item{album_id}{An album ID (\code{album_id} returned in \code{\link{get_song_meta}})}

\item{access_token}{Genius' client access token, defaults to \code{genius_token}}
}
\description{
Scrape an album's tracklisting, and song meta data, given an album ID.
}
\examples{
\dontrun{
scrape_tracklist(album_id = 337082)
}
}
