% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_artist}
\alias{search_artist}
\title{Retrieve artist identifiers for a search term}
\usage{
search_artist(search_term, n_results = 10)
}
\arguments{
\item{search_term}{A character string to search for artist matches.}

\item{n_results}{Number of results to return.}
}
\description{
The Genius API lets you search hosted content (all songs). Use \code{search_artist} to
return \code{artist_id}, \code{artist_name} and \code{artist_url} for all matches to a search.
}
\examples{
search_artist(search_term = "Kanye")
}
