% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_song}
\alias{search_song}
\title{Retrieve song identifiers for a search term}
\usage{
search_song(search_term, n_results = 10)
}
\arguments{
\item{search_term}{A character string to search for song matches.}

\item{n_results}{Number of results to return.}
}
\description{
The Genius API lets you search hosted content (all songs). Use \code{search_song} to
return \code{song_id}, \code{song_name}, \code{lyrics_url} and
\code{artist_id} for all matches to a search.
}
\examples{
search_song(search_term = "Ferrari")
}
