% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_track}
\alias{search_track}
\title{Retrieve track info for a search term}
\usage{
search_track(search_term, n_results = 10)
}
\arguments{
\item{search_term}{A character string to search for track matches.}

\item{n_results}{Number of results to return.}
}
\description{
The Genius API lets you search hosted content (all songs). Use \code{search_track} to
return \code{track_id}, \code{track_name}, \code{lyrics_url} and
\code{artist_id} for all matches to a search.
}
\examples{
search_track(search_term = "Ferrari")
}
