% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{song_meta}
\alias{song_meta}
\title{Retrieve meta data for a song}
\usage{
song_meta(song_id)
}
\arguments{
\item{song_id}{A song ID (\code{song_id} returned in \code{\link{search_song}}).}
}
\description{
The Genius API lets you search for meta data for a song, given a song ID.
}
\examples{
song_meta(song_id = 2290596)
}
