% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/songs.R
\name{tidy_custom_performances}
\alias{tidy_custom_performances}
\title{Extract custom performances from a Genius song}
\usage{
tidy_custom_performances(x)
}
\value{
a tibble
}
\description{
Extract "custom performances" (i.e. other song credits) info from a Genius song object,
as a tidy tibble.
}
\examples{
\dontrun{
song_data <- get_song(song_id = 3039923)

tidy_custom_performances(song_data)
}

}
\seealso{
See \code{\link{get_song}} to generate a Genius song object.

Other song: 
\code{\link{get_song_df}()},
\code{\link{get_song}()},
\code{\link{tidy_producer_artists}()},
\code{\link{tidy_song_relationships}()},
\code{\link{tidy_writer_artists}()}
}
\concept{song}
