% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/songs.R
\name{tidy_producer_artists}
\alias{tidy_producer_artists}
\title{Extract producer credits from a Genius song}
\usage{
tidy_producer_artists(x)
}
\value{
a tibble
}
\description{
Extract "producer artists" (i.e. producer credits) info from a Genius song object,
as a tidy tibble.
}
\examples{
\dontrun{
song_data <- get_song(song_id = 3039923)

tidy_producer_artists(song_data)
}

}
\seealso{
See \code{\link{get_song}} to generate a Genius song object.

Other song: 
\code{\link{get_song_df}()},
\code{\link{get_song}()},
\code{\link{tidy_custom_performances}()},
\code{\link{tidy_song_relationships}()},
\code{\link{tidy_writer_artists}()}
}
\concept{song}
