% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/songs.R
\name{tidy_song_relationships}
\alias{tidy_song_relationships}
\title{Extract song relationships from a Genius song}
\usage{
tidy_song_relationships(x)
}
\value{
a tibble
}
\description{
Extract "song relationships" info from a Genius song object, as a tidy tibble.
}
\examples{
\dontrun{
song_data <- get_song(song_id = 3039923)

tidy_song_relationships(song_data)
}

}
\seealso{
See \code{\link{get_song}} to generate a Genius song object.

Other song: 
\code{\link{get_song_df}()},
\code{\link{get_song}()}
}
\concept{song}
