% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/songs.R
\name{tidy_song_writers}
\alias{tidy_song_writers}
\title{Extract writer credits from a Genius song}
\usage{
tidy_song_writers(x)
}
\arguments{
\item{x}{A \code{genius_song} object}
}
\value{
a tibble
}
\description{
Extract "writer artists" (i.e. writer credits) info from a Genius song object,
as a tidy tibble.
}
\examples{
\dontrun{
song <- get_song(song_id = 3039923)

tidy_song_writers(song)
}

}
\seealso{
See \code{\link{get_song}} to generate a Genius song object.

Other song: 
\code{\link{get_song_df}()},
\code{\link{get_song}()},
\code{\link{tidy_album_performances}()},
\code{\link{tidy_song_performances}()},
\code{\link{tidy_song_producers}()},
\code{\link{tidy_song_relationships}()}
}
\concept{song}
