% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geniusr-package.R
\docType{package}
\name{geniusr-package}
\alias{geniusr}
\alias{geniusr-package}
\title{geniusr: Tools for Working with the 'Genius' API}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides tools to interact nicely with the 'Genius' API \url{https://docs.genius.com/}. Search hosted content, extract associated metadata and retrieve lyrics with ease.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ewenme.github.io/geniusr/}
  \item \url{https://github.com/ewenme/geniusr}
  \item Report bugs at \url{https://github.com/ewenme/geniusr/issues}
}

}
\author{
\strong{Maintainer}: Ewen Henderson \email{ewenhenderson@gmail.com} (\href{https://orcid.org/0000-0002-4748-4693}{ORCID})

}
\keyword{internal}
