% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{get_album_meta}
\alias{get_album_meta}
\title{Retrieve metadata for an album}
\usage{
get_album_meta(album_id, access_token = genius_token())
}
\arguments{
\item{album_id}{ID of the album (\code{album_id} within an object returned by
\code{\link{get_song}})}

\item{access_token}{Genius' client access token, defaults to \code{genius_token}}
}
\value{
a tibble
}
\description{
The Genius API lets you return data for a specific album, given an album ID.
\code{get_album_meta} returns this data in a tidy, but reduced, format.
}
\examples{
\dontrun{
get_album_meta(album_id = 337082)
}
}
\seealso{
[get_album()] to return data in full as a list.
}
