% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{get_artist_meta}
\alias{get_artist_meta}
\title{Retrieve metadata for an artist}
\usage{
get_artist_meta(artist_id, access_token = genius_token())
}
\arguments{
\item{artist_id}{ID of the artist (\code{artist_id} within an object returned by
\code{\link{search_artist}})}

\item{access_token}{Genius' client access token, defaults to \code{genius_token}}
}
\value{
a tibble
}
\description{
The Genius API lets you search for meta data for an artist, given an
artist ID. \code{get_artist_meta} returns this data in a tidy, but
reduced, format.
}
\examples{
\dontrun{
get_artist_meta(artist_id = 16751)
}
}
\seealso{
[get_artist()] to return data in full as a list.
}
