% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/songs.R
\name{get_artist_songs}
\alias{get_artist_songs}
\title{Retrieve meta data for a song}
\usage{
get_artist_songs(artist_id, include_features = FALSE,
  access_token = genius_token())
}
\arguments{
\item{artist_id}{An artist ID (\code{artist_id} returned in \code{\link{search_artist}})}

\item{include_features}{Return results where artist isn't the primary artist (logical, defaults to FALSE)}

\item{access_token}{Genius' client access token. Defaults to \code{genius_token}}
}
\description{
The Genius API lets you search for meta data for a song, given a song ID.
}
\examples{
get_artist_songs(artist_id = 1421)
}
