% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lyrics.R
\name{get_lyrics}
\alias{get_lyrics}
\title{Retrieve lyrics assoicated with a Genius song ID}
\usage{
get_lyrics(song_id)
}
\arguments{
\item{song_id}{song ID (like in \code{song_id} returned by \code{\link{search_song}})}
}
\description{
Get lyrics from Genius' lyric pages.
}
\examples{
get_lyrics(song_id = 3214267)
}
