% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lyrics.R
\name{get_lyrics}
\alias{get_lyrics}
\title{Retrieve lyrics from a Genius lyrics URL}
\usage{
get_lyrics(lyrics_url)
}
\arguments{
\item{lyrics_url}{lyric page URL (like in \code{lyrics_url} returned by \code{\link{search_track}})}
}
\description{
Get lyrics from Genius' lyric pages.
}
\examples{
get_lyrics(lyrics_url = "/Mike-will-made-it-california-rari-lyrics")
}
