% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{get_song_meta}
\alias{get_song_meta}
\title{Retrieve meta data for a song}
\usage{
get_song_meta(song_id, access_token = genius_token())
}
\arguments{
\item{song_id}{A song ID (\code{song_id} returned in \code{\link{search_song}})}

\item{access_token}{Genius' client access token, defaults to \code{genius_token}}
}
\description{
The Genius API lets you search for meta data for a song, given a song ID.
}
\examples{
\dontrun{
get_song_meta(song_id = 3039923)
}
}
