% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{print.genius_resource}
\alias{print.genius_resource}
\title{Slightly more human-readable output for genius_resource objects}
\usage{
\method{print}{genius_resource}(x, ...)
}
\arguments{
\item{x}{a genius_resource object}

\item{...}{ignored}
}
\description{
Slightly more human-readable output for genius_resource objects
}
