% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_artist}
\alias{search_artist}
\title{Search artists on Genius}
\usage{
search_artist(search_term, n_results = 10, access_token = genius_token())
}
\arguments{
\item{search_term}{A character string to search for}

\item{n_results}{Maximum no. of search results to return}

\item{access_token}{Genius' client access token, defaults to \code{genius_token}}
}
\value{
a tibble
}
\description{
The Genius API lets you search hosted content (all songs). Use \code{search_artist()}
to return \code{artist_id}, \code{artist_name} and \code{artist_url} for all unique
artist matches found using a search term.
}
\examples{
\dontrun{
search_artist(search_term = "Lil", n_results = 20)
}
}
\seealso{
Other search: 
\code{\link{search_genius}()},
\code{\link{search_song}()}
}
\concept{search}
