% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_song}
\alias{search_song}
\title{Retrieve song identifiers for associated search terms}
\usage{
search_song(search_term, n_results = 10, lyric_content_only = FALSE,
  access_token = genius_token())
}
\arguments{
\item{search_term}{A character string to search for song matches}

\item{n_results}{Maximum no. of search results to return}

\item{lyric_content_only}{Search lyric content only (i.e. ignore song title matches)}

\item{access_token}{Genius' client access token, defaults to \code{genius_token}}
}
\description{
The Genius API lets you search hosted content (all songs). Use \code{search_song} to
return \code{song_id}, \code{song_name}, \code{lyrics_url} and
\code{artist_id} for all unique song matches to a search.
}
\examples{
\dontrun{
search_song(search_term = "Gucci", n_results=50)
}
}
